requirements.txt
python-dotenv~=1.0.0
python-telegram-bot[webhooks]~=20.2.0
pymongo~=4.3.3